backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
email-validator>=2.0.0    # ← agrega esto
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings>=2.0.3
